function that returns the length of a string

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -12,7 +12,7 @@
  *
  * Return: The length of @str.
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 return (strlen(s));
 }
